requirements.txt
tensorflow==2.12.1
keras==2.12.0
matplotlib
numpy
pandas
Pillow
scikit_learn
scipy
tabulate
ipykernel
kaggle
PyYAML